Cleaned file structure, add Links, Switches, Routers, etc
