+++ public/styles.css
@@ -5,7 +5,7 @@
 .settings,
 .quiz,
 .results {
-    margin-top: 150px;
+    margin-top: 50px;
 }
 
 .margin-top {
