+++ src/Frontend/Components/QuizPage/SubmitQuiz.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { Link } from "react-router-dom";
 
 interface Props {
     submitQuiz: any;
@@ -8,12 +9,14 @@ function SubmitQuiz(props: Props) {
     const { submitQuiz } = props;
 
     return (
-        <button
-            className="btn btn-lg text-uppercase btn-success mt-3 mb-5"
-            onClick={submitQuiz}
-        >
-            Submit Quiz!
-        </button>
+        <Link to="/quiz/results">
+            <button
+                className="btn btn-lg text-uppercase btn-success mt-3 mb-5"
+                onClick={submitQuiz}
+            >
+                Submit Quiz!
+            </button>
+        </Link>
     );
 }
