+++ src/Frontend/Components/SettingsPage/SettingsPage.tsx
@@ -1,7 +1,6 @@
 import React from "react";
-import Navbar from "../QuizPage/Navbar";
-import Description from "./Description";
-import Settings from "./Settings";
+import Description from "./SettingsComponents/Description";
+import Settings from "./SettingsComponents/Settings";
 
 interface Props {
     startGame: any;
@@ -11,12 +10,9 @@ function SettingsPage(props: Props) {
     const { startGame } = props;
 
     return (
-        <div className="container-fluid settings-page">
-            <Navbar message={"Quiz App"} timer={""} />
-            <div className="container margin-top pt-5">
-                <Description />
-                <Settings startGame={startGame} />
-            </div>
+        <div className="container settings">
+            <Description />
+            <Settings startGame={startGame} />
         </div>
     );
 }
